Inject embed script earlier
Waiting for the wombat banner callback is too long. The embed script
needs to be injected ASAP so that it takes precedence over any embed
script included by the page. Otherwise, the page might embed Hypothesis
itself, which is problematic because such an embedding would be subject
to the rewriter and the result is a much degraded experience (slowness,
unable to log in due to cookie problems, etc).