Block content from a prefix blacklist
We need to be able to omit content from the via proxy if content owners
request it. This is currently implemented using some rather dubious
nginx rules, and needs to move into this application for the move to
Skyliner.

This commit implements a WSGI middleware responsible for serving a
static page for blocked content, and adds the blocked content list to
the repository in a flat file.

I've also rearranged the middleware stack so that:

- we cannot accidentally block our own static content
- we do not bother doing header sanitisation on our own block page or
  on static content